#5 Added confirmation dialog for grammar deletion.

+++ Assets/Michelangelo/Model/Grammar.cs
@@ -66,7 +66,7 @@ namespace Michelangelo.Model {
             if (showInstantiate && GUILayout.Button("Instantiate")) {
                 MichelangeloSession.InstantiateGrammar(id).Then(onResolved).Catch(onRejected);
             }
-            if (isOwner && GUILayout.Button("Delete")) {
+            if (isOwner && GUILayout.Button("Delete") && EditorUtility.DisplayDialog("Delete grammar?", $"Are you sure you want to delete grammar \"{name}\"?", "Delete", "Cancel")) {
                 MichelangeloSession.DeleteGrammar(id).Then(onResolved).Catch(onRejected);
             }
             EditorGUILayout.EndHorizontal();
